+++ sheeter/tmpls/tmplGo.go
@@ -140,7 +140,6 @@ func (this *Sheeter) FromData() bool {
 	} // if
 
 	waitGroup := sync.WaitGroup{}
-	waitGroup.Add({{len $.Alone}} + {{len $.Merge}})
 	result := atomic.Bool{}
 	result.Store(true)
 
@@ -149,28 +148,23 @@ func (this *Sheeter) FromData() bool {
 		&this.{{.StructName}},
 {{- end}}
 	} {
-		tmpl := itor
-
-		go func() {
-			defer waitGroup.Done()
-			filename := tmpl.FileName()
+		waitGroup.Go(func() {
+			filename := itor.FileName()
 			data := this.loader.Load(filename)
 
 			if len(data) == 0 {
 				return
 			} // if
 
-			if err := tmpl.FromData(data, true, this.progress); err != nil {
+			if err := itor.FromData(data, true, this.progress); err != nil {
 				this.loader.Error(filename.File(), err)
 				result.Store(false)
 			} // if
-		}()
+		})
 	} // for
 {{- range $.Merge}}
 
-	go func() {
-		defer waitGroup.Done()
-
+	waitGroup.Go(func() {
 		for i, itor := range []Reader{
 {{- range $name := .MemberName}}
 			&this.{{$name}},
@@ -188,7 +182,7 @@ func (this *Sheeter) FromData() bool {
 				result.Store(false)
 			} // if
 		} // for
-	}()
+	})
 {{- end}}
 
 	waitGroup.Wait()