+++ sheeter/utils/term.go
@@ -21,7 +21,7 @@ func (this MergeTerm) Name() string {
 // Member 取得合併成員
 func (this MergeTerm) Member() (result []SheetTerm) {
 	if _, member, ok := strings.Cut(string(this), sheeter.TokenName); ok {
-		for _, itor := range strings.Split(member, sheeter.TokenTerm) {
+		for itor := range strings.SplitSeq(member, sheeter.TokenTerm) {
 			result = append(result, SheetTerm(itor))
 		} // for
 	} // if