+++ support/example/sheet/codeGo/sheeter.go
@@ -32,29 +32,25 @@ func (this *Sheeter) FromData() bool {
 	} // if
 
 	waitGroup := sync.WaitGroup{}
-	waitGroup.Add(1 + 0)
 	result := atomic.Bool{}
 	result.Store(true)
 
 	for _, itor := range []Reader{
 		&this.ExampleData,
 	} {
-		tmpl := itor
-
-		go func() {
-			defer waitGroup.Done()
-			filename := tmpl.FileName()
+		waitGroup.Go(func() {
+			filename := itor.FileName()
 			data := this.loader.Load(filename)
 
 			if len(data) == 0 {
 				return
 			} // if
 
-			if err := tmpl.FromData(data, true, this.progress); err != nil {
+			if err := itor.FromData(data, true, this.progress); err != nil {
 				this.loader.Error(filename.File(), err)
 				result.Store(false)
 			} // if
-		}()
+		})
 	} // for
 
 	waitGroup.Wait()