+++ support/prototype/sheeter.go
@@ -32,7 +32,6 @@ func (this *Sheeter) FromData() bool {
 	} // if
 
 	waitGroup := sync.WaitGroup{}
-	waitGroup.Add(4)
 	result := atomic.Bool{}
 	result.Store(true)
 
@@ -40,27 +39,22 @@ func (this *Sheeter) FromData() bool {
 		&this.Alone0,
 		&this.Alone1,
 	} {
-		tmpl := itor
-
-		go func() {
-			defer waitGroup.Done()
-			filename := tmpl.FileName()
+		waitGroup.Go(func() {
+			filename := itor.FileName()
 			data := this.loader.Load(filename)
 
 			if len(data) == 0 {
 				return
 			} // if
 
-			if err := tmpl.FromData(data, true, this.progress); err != nil {
+			if err := itor.FromData(data, true, this.progress); err != nil {
 				this.loader.Error(filename.File(), err)
 				result.Store(false)
 			} // if
-		}()
+		})
 	} // for
 
-	go func() {
-		defer waitGroup.Done()
-
+	waitGroup.Go(func() {
 		for i, itor := range []Reader{
 			&this.Alone0,
 			&this.Alone1,
@@ -77,11 +71,9 @@ func (this *Sheeter) FromData() bool {
 				result.Store(false)
 			} // if
 		} // for
-	}()
+	})
 
-	go func() {
-		defer waitGroup.Done()
-
+	waitGroup.Go(func() {
 		for i, itor := range []Reader{
 			&this.Alone0,
 			&this.Alone1,
@@ -98,7 +90,7 @@ func (this *Sheeter) FromData() bool {
 				result.Store(false)
 			} // if
 		} // for
-	}()
+	})
 
 	waitGroup.Wait()
 	this.progress.Complete()