+++ support/verify/codeGo/sheeter.go
@@ -33,34 +33,28 @@ func (this *Sheeter) FromData() bool {
 	} // if
 
 	waitGroup := sync.WaitGroup{}
-	waitGroup.Add(1 + 1)
 	result := atomic.Bool{}
 	result.Store(true)
 
 	for _, itor := range []Reader{
 		&this.VerifyData,
 	} {
-		tmpl := itor
-
-		go func() {
-			defer waitGroup.Done()
-			filename := tmpl.FileName()
+		waitGroup.Go(func() {
+			filename := itor.FileName()
 			data := this.loader.Load(filename)
 
 			if len(data) == 0 {
 				return
 			} // if
 
-			if err := tmpl.FromData(data, true, this.progress); err != nil {
+			if err := itor.FromData(data, true, this.progress); err != nil {
 				this.loader.Error(filename.File(), err)
 				result.Store(false)
 			} // if
-		}()
+		})
 	} // for
 
-	go func() {
-		defer waitGroup.Done()
-
+	waitGroup.Go(func() {
 		for i, itor := range []Reader{
 			&this.VerifyData,
 		} {
@@ -76,7 +70,7 @@ func (this *Sheeter) FromData() bool {
 				result.Store(false)
 			} // if
 		} // for
-	}()
+	})
 
 	waitGroup.Wait()
 	this.progress.Complete()